Chapter 9 NSUndoManager A challenge to get to this point, being able to undo all additions and removals of items to the employee list. It is working well now. Things to overcome were Xamarin.mac/MonoMac not properly implementing the NSUndoManager. Specifically this cannot be done in Xamarin.Mac:
Obj-C:
NSUndoManager undo = [self undoManager];
[[undo prepareWithInvocationTarget:self] removeObjectFromEmployeesAtIndex:index];

and

NSUndoManager undo = [self undoManager];
[[undo prepareWithInvocationTarget:self] insertObject:person inEmployeesAtIndex:index];

C#:
NSUndoManager undo = this.UndoManager;
// Call that throws exception. Bug report: https://bugzilla.xamarin.com/show_bug.cgi?id=19623
undo.PrepareWithInvocationTarget(this);

Even if the above didn't crash, I still would not know how to then call the
appropriate method as in the Obj-C above, let alone pass in the index and the Person object.

I was able to use RegisterUndoWithTarget to get the methods called but could not figure out how to pass the index and the Person object to 
the undo manager so they could be used to undo and redo actions.

I worked around this by managing my own index and Person object stacks to keep a reference of the arguments I could not pass to the undo and redo methods.

The other thing to overcome was checking the status of the undo manager, See the code.